+++ src/about.js
@@ -4,3 +4,4 @@ import "../node_modules/@glidejs/glide/dist/css/glide.core.min.css";
 import "../node_modules/@glidejs/glide/dist/css/glide.theme.min.css";
 import "./js-modules/github-api-module.js";
 import "./js-modules/commits-module.js";
+import "./js-modules/commits-list-module.js";
