+++ src/index.js
@@ -6,3 +6,4 @@ import "./js-modules/cards-module.js";
 import "./js-modules/loader-module.js";
 import "./js-modules/time-module.js";
 import "./js-modules/validation-module.js";
+import "./js-modules/card-list-module.js";
