+++ src/js-modules/github-api-module.js
@@ -14,9 +14,10 @@ class GitHubApi {
         
           .then(res => {
             if (res.ok) {
-                return res.json();
-            }
+              return res.json();
+            } else {
             return Promise.reject(`Ошибка: ${res.status}`);
+            }
           })
     }
 }
