--- src/searcher/__search/searcher__search.css
@@ -1,18 +1,0 @@
-.searcher__search {
-    max-width: 608px;
-    margin: 0 auto;
-}
-
-@media screen and (max-width: 769px) {
-
-    .searcher__search {
-        max-width: 448px;
-    }
-}
-
-@media screen and (max-width: 480px) {
-
-    .searcher__search {
-        margin-left: 16px;
-    }
-}
