--- src/searcher/__subtitle/searcher__subtitle.css
@@ -1,26 +1,0 @@
-.searcher__subtitle {
-    max-width: 580px;
-    font-family: Roboto;
-    font-weight: normal;
-    font-size: 18px;
-    line-height: 24px;
-    color: white;
-    margin-top: 32px;
-    margin-bottom: 0;
-}
-
-@media screen and (max-width: 769px) {
-
-    .searcher__subtitle {
-        margin-top: 16px;
-    }
-}
-
-@media screen and (max-width: 380px) {
-
-    .searcher__subtitle {
-        font-size: 20px;
-        line-height: 26px;
-        margin-top: 56px;
-    }
-}
