--- src/searcher/__title/searcher__title.css
@@ -1,27 +1,0 @@
-.searcher__title {
-    max-width: 590px;
-    font-family: Roboto Slab;
-    font-size: 60px;
-    line-height: 64px;
-    color: white;
-    font-weight: lighter;
-    margin-top: 80px;
-    margin-bottom: 0;
-}
-
-@media screen and (max-width: 769px) {
-
-    .searcher__title {
-        font-size: 36px;
-        line-height: 40px;
-        margin-top: 40px;
-
-    }
-}
-
-@media screen and (max-width: 380px) {
-
-    .searcher__title {
-        margin-top: 30px;
-    }
-}
