--- src/searcher/searcher.css
@@ -1,26 +1,0 @@
-@import url('./__form/searcher__form.css');
-@import url('./__search/searcher__search.css');
-@import url('./__subtitle/searcher__subtitle.css');
-@import url('./__title/searcher__title.css');
-
-.searcher {
-    width: 100%;
-    min-height: 576px;
-    background-image: url('../../images/georgia-de-lotz--UsJoNxLaNo-unsplash.png');
-    background-size: cover;
-    background-position: center;
-}
-
-@media screen and (max-width: 769px) {
-
-    .searcher {
-        height: 400px;
-    }
-}
-
-@media screen and (max-width: 380px) {
-
-    .searcher {
-        height: 536px;
-    }
-}
